refactor: marshal utils to /internal
