+++ packages/vats/test/bootstrapTests/supports.js
@@ -9,7 +9,7 @@ import childProcessAmbient from 'child_process';
 import { Fail } from '@agoric/assert';
 import { buildSwingset } from '@agoric/cosmic-swingset/src/launch-chain.js';
 import { BridgeId, makeTracer, VBankAccount } from '@agoric/internal';
-import { unmarshalFromVstorage } from '@agoric/internal/src/lib-chainStorage.js';
+import { unmarshalFromVstorage } from '@agoric/internal/src/marshal.js';
 import { makeFakeStorageKit } from '@agoric/internal/src/storage-test-utils.js';
 import { initSwingStore } from '@agoric/swing-store';
 import { kunser } from '@agoric/swingset-liveslots/test/kmarshal.js';
